+++ UI/ColorVision.Database/DatabaseType.cs
@@ -1,8 +1,0 @@
-namespace ColorVision.Database
-{
-    public enum DatabaseType
-    {
-        MySql,
-        Sqlite
-    }
-}